JAVA-1761: Add OSGi descriptors
Motivation:

To enable the driver to be used in OSGi containers.

Modifications:

Use maven-bundle-plugin to add OSGi information to the generated
MANIFEST.MF.

Add jctools as a dependency instead of using netty-shaded
MpscLinkedAtomicQueue in DefaultWriteCoalescer.  This was required
because netty does not export its shaded jctools packages in its
Export-Package definition.

Change Reflection.buildFromConfig to use the expectedSuperType's
ClassLoader instead of the current Thread's ClassLoader when loading a
class.  This enables the driver to load classes from the core bundle in
OSGi, but may be problematic in other situations so this may be an
incomplete solution.

Add OSGi integration tests.

Make all dependency versions properties so they may be passed to
failsafe plugin for use in OSGi tests.

Result:

The driver now supports OSGi.