Optional coalscing of schema refresh queries
By default the driver has the ability to coalesce schem refresh queries
to try and reduce the number of queries sent. In case a view and a table
refresh is needed a refresh for the keyspace would be generated.

While in most cases thisoptimization is helpfull - in case there are
100s of tables in a single keyspace this will cause fetching all
keyspace information which will cause extra load on the system (in
scylla's case a single shard).

Add the ability to disable this by setting the
setMaxPendingRefreshSchemaRequests to 1

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>

Update

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>